requirements.txt
pyserial==3.4
cryptography==37.0.2
flaky==3.7.0
google-api-python-client==2.49.0
google-auth-httplib2==0.1.0
google-auth==2.6.6
google-cloud-iot==2.5.0
google-cloud-pubsub==2.12.1
google-cloud-compute==1.3.1
google-cloud-bigquery
google-cloud-functions
google-cloud-storage
google-cloud-logging==3.0.0
paho-mqtt==1.6.1
pyjwt==2.4.0
rfc3339==6.0
pandas
db-dtypes

